GRAILS-11989 - The trait has moved
See https://github.com/grails/grails-core/commit/480b361517794a3218e2f4ba317f093fec36798d#diff-40c8965de787b24de5800319e1fc62ed
